Refactor multiple trowel item classes into a single TrowelItem class
